+++ OtterLib/core/include/Orchestrator.hpp
@@ -1,28 +1,35 @@
-
 #ifndef ORCHESTRATOR_H
 #define ORCHESTRATOR_H
+
 #include "registry.hpp"
+
+
+
 namespace Core {
-class orchestrator {
+class Orchestrator {
   public:
-    orchestrator() : _components(), _entity() {}
+    Orchestrator() : _components(), _entity() {}
 
-    Entity CreateEntity();
+  Entity CreateEntity() {return _entity.CreateEntity();}
+  
+  template <class C> typename ComponentManager::container_t<C> &register_component() {return _components.register_component<C>();}
+  template <class C> typename ComponentManager::container_t<C> &get_components() {return _components.get_components<C>();}
+  template <class C> typename ComponentManager::container_t<C> const &get_components() const {return _components.get_components<C>();}
+  template <typename C>
+  typename ComponentManager::reference_type<C> add_component(Entity const &addr, C &&component) {
+    return _components.add_component<C>(addr, component);
+  }
 
-    template <class C> container_t<C> &register_component() {}
-    // template <class C> container_t<C> &get_components() {}
-    //   template <class C> container_t<C> const &get_components() const {}
-    // template <typename C>
-    // eference_type<C> add_component(Entity const &addr, C &&component) {}
+  template <typename C, typename... Params>
+  typename ComponentManager::reference_type<C> emplace_component(Entity const &addr, Params &&...var) {
+    return emplace_component<C>(addr, var...);
+  }
 
-    // template <typename C, typename... Params>
-    //  reference_type<C> emplace_component(Entity const &addr, Params &&...var) {}
-
-    //    template <typename C> void remove_component(Entity const &addr) {}
-    void remove_entity(Entity const &addr) {
-        _components.remove_entity(addr);
-        _entity.destroyEntity(addr);
-    }
+  template <typename C> void remove_component(Entity const &addr) {}
+  void remove_entity(Entity const &addr) {
+    _components.remove_entity(addr);
+    _entity.destroyEntity(addr);
+  }
 
   private:
     ComponentManager _components;
