+++ OtterLib/core/include/OtterCore.hpp
@@ -1,3 +1,23 @@
 #pragma once
 
 #include "Orchestrator.hpp"
+
+namespace Core {
+/*
+class OtterCore {
+  OtterCore() : _data() {
+  }
+  void engineStart();
+
+
+private:
+  void loop();
+  void init();
+
+  Orchestrator _data;
+  //main
+
+}
+*/
+
+}
