+++ OtterLib/core/src/registry.cpp
@@ -1,4 +1,3 @@
 
 
 #include "OtterCore.hpp"
-int main() {}
